+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Fixed issue with OpenMetrics counter suffixes for Prometheus
-  ([#5623](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5623))
+  ([#5646](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5646))
 
 ## 1.9.0-alpha.1