+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Fixed issue with OpenMetrics counter suffixes for Prometheus
+* Fixed issue with OpenMetrics suffixes for Prometheus
   ([#5646](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5646))
 
 ## 1.9.0-alpha.1